requirements.txt
flask
pillow
opencv-python
torch
torchvision